+++ src/pycatdap/__init__.py
@@ -26,7 +26,7 @@ from pycatdap._version import (
 from pycatdap.catdap1 import Catdap1Result, catdap1
 from pycatdap.catdap2 import Catdap2Result, catdap2
 from pycatdap.eda import DescribeResult, describe
-from pycatdap.plot import plot_missing, plot_target, plot_variable
+from pycatdap.plot import plot_missing, plot_pair, plot_target, plot_variable
 
 __all__ = [
     "__version__",
@@ -41,6 +41,7 @@ __all__ = [
     "datasets",
     "describe",
     "plot_missing",
+    "plot_pair",
     "plot_target",
     "plot_variable",
     "target_summary",